more precision

+++ csrc/multi_tensor_lamb.cu
@@ -297,7 +297,7 @@ struct LAMBStage2Functor
           MATH_T old_p = r_p[ii];
 	  r_p[ii] = static_cast<MATH_T>(r_p[ii]) - (ratio * static_cast<MATH_T>(r_update[ii]));
 	  if (true) { //tensor_loc== 1 && i_start == 0 && ii == 0) {
-            printf("tensor_loc:%d,tensor_num:%d,old_p:%.16f,ratio:%.16f,update:%.16f,new_p:%.16f\n", tensor_loc, tensor_num, old_p, ratio, r_update[ii], r_p[ii]);
+            printf("tensor_loc:%d,tensor_num:%d,old_p:%.16f,ratio:%.16f,update:%.16f,new_p:%.16f\n", tensor_loc, tensor_num, old_p, ratio, (float) r_update[ii], (float) r_p[ii]);
           }
         }
         load_store(p, r_p, i_start, 0);
